refactor: remove unused codes, move core/agent module into dataset retrieval feature
