--- api/core/tool/provider/errors.py
@@ -1,2 +1,0 @@
-class ToolValidateFailedError(Exception):
-    description = "Tool Provider Validate failed"
